#45 :recycle: refactor: header의 예약건 수 + 방문일자 텍스트 렌더링은 ReservationCard의 역할이 아니라고 판단하여 Detail 컴포넌트에 위임
